requirements.txt
Appium-Python-Client==4.2.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
psutil==6.0.0
requests==2.32.3
selenium==4.25.0
pillow==10.4.0
opencv-python==4.10.0.84
